Refaktorering og lagt til flere tester

+++ src/main/kotlin/no/nav/bidrag/behandling/dto/v1/beregning/ResultatForskuddsberegning.kt
@@ -19,7 +19,7 @@ data class ResultatBeregningBarnDto(
         val regel: String,
         val sivilstand: Sivilstandskode?,
         val inntekt: BigDecimal,
-        val antallBarnIHusstanden: Double,
+        val antallBarnIHusstanden: Int,
     )
 }
